Update libprov and use its cmake/provider.cmake
libprov/cmake/provider.cmake provides a macro 'build_provider' with a
standard method to build a provider.

With this, CMakeLists.txt is simplified quite a lot.

Additionally, 'enable_testing()' is replaced with 'include(CTest)'.
It doesn't make much different for the moment, but enables features that may
be used later on.