"extra" provider license change
I finally decided to change the license to the Lesser GNU Public License 3.0
or later.  This allows this provider to be used with and runtime linked with
any application, while assuring that the provider code itself remains free
according to [the four freedoms].

I chose to use the [REUSE] tool for all copyright and license annotations,
and encourage everyone to use that same format (SPDX) and tool (optional,
but makes life a little easier).

[the four freedoms]: https://www.gnu.org/philosophy/free-sw.html.en

Fixes #3